Add oauth test
