+++ tests/oauth/pyproject.toml
@@ -1,0 +1,13 @@
+[project]
+name = "oauth"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "dotenv>=0.9.9",
+    "microsoft-teams-app",
+]
+
+[tool.uv.sources]
+microsoft-teams-app = { workspace = true }
